tests/protocol: Skip these tests if default attach-method is not 'appliance'. (cherry picked from commit bc917a0efc03b8c28966c9ab474f525752ebe440)

+++ tests/protocol/test-qemudie-synch.sh
@@ -20,6 +20,11 @@
 
 set -e
 
+if [ "$(../../fish/guestfish get-attach-method)" != "appliance" ]; then
+    echo "$0: test skipped because default attach-method is not 'appliance'"
+    exit 77
+fi
+
 rm -f test.pid test1.img
 
 ../../fish/guestfish -N disk <<'EOF'
